pyproject.toml
[build-system]
requires = ["setuptools>=40.8.0", "wheel"]
requires-python = "==3.11"
build-backend = "setuptools.build_meta"

[project]
name = "chameleon"
version = "1.0.0"
description = "Chameleon: Mixed-Modal Early-Fusion Foundation Models"
readme = "README.md"
authors = [{name = "Chameleon Team"}]
license = {file = "LICENSE"}
dependencies = [
    "transformers==4.36.2",
    "xformers==0.0.23",
    "Pillow==10.0.*",
    "click",
    "flask",
    "flask_socketio",
]

[tool.setuptools.package-data]
chameleon = ["miniviewer/*.html"]

[tool.setuptools.packages.find]
include = ["chameleon*"]

[tool.mypy]
python_version = 3.11
disallow_untyped_calls = true
disallow_any_explicit = true
disallow_any_generics = true
ignore_missing_imports = true
follow_imports = 'skip'

[tool.isort]
profile = "black"

[tool.ruff]
extend-select = ["I"]
